+++ src/termui/__init__.py
@@ -1,6 +1,6 @@
 from .app import App
 from .screen import Screen
-from .layout.div import Div
+from .layouts.div import Div
 from .input import InputHandler, Keybind, keybind
 
 __all__ = ["App", "Screen", "Div", "InputHandler", "Keybind", "keybind"]
